+++ e-library/src/Components/borrow/Borrowbook.jsx
@@ -2,37 +2,37 @@ import React from 'react';
 import useBorrowBook from '../hooks/borrowBook';
 
 const GetBookbyId = () => {
-  const { bookid, selectbook } = useBorrowBook();
+  const { book, selectbook } = useBorrowBook();
   return (
     <>
       <div className="container2">
         <div className="img">
-          <img src={bookid.imageUrl} alt={bookid.title} />
+          <img src={book.imageUrl} alt={book.title} />
           <button type="button" onClick={selectbook}>BORROW</button>
         </div>
       </div>
       <div className="container3">
         <p>
           Title:
-          {bookid.title}
+          {book.title}
         </p>
       </div>
       <div className="container3">
         <p>
           About:
-          {bookid.description}
+          {book.description}
         </p>
       </div>
       <div className="container3">
         <p>
           Price:
-          {bookid.price}
+          {book.price}
         </p>
       </div>
       <div className="container3">
         <p>
           Owner:
-          {bookid.author}
+          {book.author}
         </p>
       </div>
     </>
